separate controller for registration
